http://issues.apache.org/activemq/browse/AMQCPP-103
Fixing some compiler warnings on the decaf code while re-inserting it into the amqcpp baseline.

git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@652472 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/decaf/nio/CharBuffer.cpp
@@ -428,7 +428,7 @@ int CharBuffer::compareTo( const CharBuffer& value ) const {
         compareRemaining--;
     }
 
-    return remaining() - value.remaining();
+    return (int)( remaining() - value.remaining() );
 }
 
 ////////////////////////////////////////////////////////////////////////////////
