+++ src/main/decaf/nio/IntBuffer.cpp
@@ -267,7 +267,7 @@ int IntBuffer::compareTo( const IntBuffer& value ) const {
         compareRemaining--;
     }
 
-    return remaining() - value.remaining();
+    return (int)( remaining() - value.remaining() );
 }
 
 ////////////////////////////////////////////////////////////////////////////////
