+++ src/main/decaf/nio/LongBuffer.cpp
@@ -267,7 +267,7 @@ int LongBuffer::compareTo( const LongBuffer& value ) const {
         compareRemaining--;
     }
 
-    return remaining() - value.remaining();
+    return (int)( remaining() - value.remaining() );
 }
 
 ////////////////////////////////////////////////////////////////////////////////
